Extract TrelloList into trello_list.dart
