+++ lib/src/sdk/lists/lists.dart
@@ -1,2 +1,3 @@
 export 'list_client.dart';
 export 'list_models.dart';
+export 'trello_list.dart';
